api/requirements.txt
Flask
mysql-connector
flask-cors
gunicorn